JS: Utilize arrow functions to avoid ambiguity with `this` reference
Unlike normal functions on javascript, the arrow functions don't bind their own
scope but inherit the parent one. Hence `this` is the same `this` within the
callback defined with an arrow function than in the code which created that
callback. By utilizing this, we don't need to create a separate `self`
reference to the closure, and the code remains a bit easier to follow.