Help PHPStan to understand the class_alias
PHPStan cannot pick up any class_alias definitions from the code to be
analyzed and was producing a lot of errors. Giving the alias in a specific
bootstrap file configured for PHPStan solves the issue.

Also, PHPStan was complaining about an unnecessary sanity check in
OldNextcloudMapper. This was set to be ignored because the check was
pointless only if you believe that the type hints given as PHPDoc can never
be broken.