Define background jobs in less ancient way to be compatible with NC25+
We used to register our background tasks using the legacy adapter
OC\BackgroundJob\Legacy\RegularJob. This had been the "legacy" way
already since 10 years ago. The more modern alternative is to make the
background job classes extend suitable base class and register them
directly instead of using the adapter. NC25 has finally removed the
legacy adapters, and for this reason, our background jobs were not
running on NC25 and later.

However, also the new alternative taken into use here has been
deprecated already in Nextcloud 24; the preferred alternative would now
be to extend a new base class OCP\BackgroundJob\TimedJob instead of the
private OC\BackgroundJob\TimedJob. Naturally, the new base class is not
available on ownCloud and we have to stick with the deprecated
alternative. This will likely cause more work on some future release of
NC.

Also, the registration of the background jobs now happens within a
post-migration repair step instead of doing it again on each HTTP
request. This is the more proper way and should be better for the
overall runtime performance. It would be possible to register them also
within info.xml (at least on some cloud versions) but that wouldn't
allow us to deregister the legacy jobs.

refs https://github.com/owncloud/music/issues/1044