Fix nav pane divider lines being invisible with some themes on NC25+
These divider lines used to have a hard-coded light gray color. This
was fine previously when the navigation pane background was always
either white or black. However, on Nextcloud 25 and later, the
navigation pane is semi transparent and affected by the underlying
background picture. Hence, it often ends up being some shade of gray
and this could be very close to the color of the divider line.

This was fixed by using the main font color with some transparency. The
themes and background pictures should always be designed so that the
main font remains visible on the navigation pane.