Allow UTF-8 encoding also on playlists with the extension .m3u
Previously, playlist files with the .m3u extension were always interpreted as
ISO-8859-1 unless otherwise specified with the tag #EXTENC. However, it's
nowadays increasingly common to use UTF-8 encoding also in those files and the
use of the #EXTENC tag is uncommon.

With this commit, any .m3u file looking like valid UTF-8 is interpreted as
UTF-8. ISO-8859-1 is used only if the UTF-8 interpretation fails or the
#EXTENC tag rules so.

refs https://github.com/owncloud/music/issues/1047