Subsonic: Fix handling of the `time` argument on the `scrobble` method
According to the Subsonic API specification, the optional `time`
argument on the method `scrobble` is given as milliseconds since epoch.
However, our logic assumed it to be in seconds. Depending on the
database system in use, this caused either the scrobbling to fail or be
recorded with a time far in the future.

refs https://github.com/owncloud/music/issues/1066