Fix empty player bar shown on Nextcloud when "Show in Files" selected
When clicking the "Show in Files" link on the track details pane, there used
to be an empty embedded player bar being shown in the bottom of the Files view
on Nextcloud. This has happened at least since Nextcloud 23 but not from the
beginning of the time. The exact NC version where this has started was not
investigated, though. Also, the problem did not affect ownCloud.

This happened because NC tried to automatically launch the file viewing action
in this case but the view was not ready yet for our embedded player to work as
intended. It was fixed by not showing the player at all if attempted too early.