Subsonic: Fix unhandled exception on `deleteBookmark`
There was an unhandled exception when the Subsonic action
`deleteBookmark` was called for a non-existent bookmark, and this got
logged as an error. This happened because, by convention, the business
layer modules should throw BusinessLayerException in case the requested
item doesn't exist. The SubsonicMiddleware catches such exceptions and
emits the proper Subsonic error response. But BookmarkBusinessLayer was
throwing wrong kind of exception here and it leaked through.

refs https://github.com/owncloud/music/issues/1071