Refactoring: Declare API methods in SubsonicController as `protected`
Declare the methods forming the Subsonic API as `protected` instead of
`private`. These methods are called via reflection, by matching the
requested URL to the method name.

Essentially, this is abusing the `protected` keyword since the class
SubsonicController is not really intended for inheritance. However,
when viewing the code, this helps in separating the API methods from
truly internal helper methods. Also, this suppresses the IDE warning
about unused methods (as the IDE doesn't understand reflection).
Finally, this is the same approach which had been taken on the
AmpacheController class.