Drop the cookies to localStorage migration logic
The UI user settings like volume, shuffle, etc. have been stored to the
HTML5 localStorage since Music v1.4.0. Until now, we have had a logic
which checks the existence of these settings in the legacy cookie
format, and moves any found settings to the localStorage. It's unlikely
that there would be many people left updating from so outdated versions
anymore. And even in case that happens, the only thing happening is
that the Music app will forget those UI settings after the update, and
the user has to adjust them again.

By dropping the logic, we could get rid of the 3rd party js-cookie
library.