Simplify the structure of the Random view
Most of the complexity there was copy-paste from the All tracks view.
But the thing is, the All tracks view needs this because it may
potentially contain tens of thousands of tracks. Meanwhile, the Random
view can only contain up to 300 tracks and much simpler structure is
sufficient.

Furthermore, the bucket logic of the All tracks view has been designed
for the case where the tracks are ordered alphabetically. It made
little sense on the randomly ordered list.