Don't update the smartlist filters on enter key
This didn't work quite logically on the chosen-js fields because they
seem to leak the enter key event even when they use the event to select
an item from the list. For consistency, the enter-key handling from the
filters pane was removed altogether. It wasn't very useful in the first
place.