Optimize away the field `artistName` from the track of `collection.json`
As the `collection.json` nowadays contains artist entries also for
those artists with no own albums, it was no longer necessary to include
the name of the performing artist on each track entry. Instead, we now
use the `artistId` on the track entry to find the corresponding artist
and its name. With this change, the size of `collection.json` was
reduced by ~20% on the test system.