Fix the alphabet navigation in the Albums view after a7c8f9ff9
The navigation didn't work because the array of artists given to the
alphabetNavigation directive contained also the artists with no albums
and no corresponding element in the Albums view.

To fix this, a clear separation was made between the "collection"
containing the tree structure of the Albums view and "all artists"
containing all known artists, including the ones with no albums.