Support dragging songs/groups on the "+ New Playlist" item on the navi pane
With this action, it is now possible to create new playlist containing the
given tracks. It is possible also to drag tracks/groups one-after-another on
the "+ New Playlist" item, and all of them will be gathered on the newly
created list once the commit button is pressed. A badge replacing the "+" icon
tells, how many tracks are about to be added on the newly created list.