Avoid overflow on numeric DB fields while scanning audio files
There are a few integer fields stored to the database which get their
values from the data scanned from the input files: disk, number,
year, length, bitrate.

As we have no control over what kind of data may be scanned from the
files, the values need to be sanitized before entering them to the DB.
If the value wouldn't fit in the numeric limits of the DB field in
question, there would be an unhandled exception and that would break
the scanning process.

refs https://github.com/owncloud/music/issues/1073