Ampache: Refactoring
Got rid of the AmpacheUser utility class. This was used to share the
information about the authorized user between AmpacheMiddleware and
AmpacheController. Now we use the AmpacheSession for the same purpose.
This has the added benefit of sharing the current token, too, and we no
longer need to pass it around as an argument of dozens of functions.
The AmpacheSession also carries the information about the requested API
version which will come in handy very soon.