Ampache: Don't "unwrap" JSON responses when the session API version is >4
In the APIv4, most JSON responses had an anonymous array as the root
element. This has changed in APIv5 so that the root is now an object
with a single named property and array as its value. This is more
straight forward for us, and similar to how the Subsonic API has always
worked.