Ampache: Add all the "genre" actions
Added actions  `genres`, `genre`, `genre_artists`, `genre_albums`,
`genre_songs`. These are APIv5 replacements for the former "tag"
functions. We now provide then regardless of the API version. They are
otherwise identical to the "tag" functions but the the result elements
have the name "genre" instead of "tag".

Also, the `handshake` result as well as any results containing songs,
albums, or artists now use the key "genre" instead of "tag" when the
active API version is 5 or greater.