Ampache: Very limited support for the APIv6
There are only two things which we do differently on APIv6 compared to
APIv5 and both concern only the JSON API:
- The value `flag` on entities is shown as boolean instead of integer
- The result of action `album` is formatted consistently with all the
  other "singular" actions.
The APIv6 specification also brings quite a few new functions but we
don't currently support most of these.

Note that the Util::intCastArrayValues function, which we now took in
use, didn't previously work at all because it didn't take its array
variable as a reference. This hadn't been noticed as the function has
not been in use for a long time.

As a not-quite-related change, the (dummy) property `stream` was renamed
to`live_stream` on the genre results. This change has occurred in the
API spec between v4 and v5 (i.e. when moving from "tags" to "genres")
but we apply it now regardless of the version. It's unlikely that any
client has utilized this property previously for anything, given that
the live_stream support came to the API only on v5 and it really isn't
integrated in any way with the genres even in v6 (e.g. there is no
"genre_streams" action).