Add nop getNowPlaying route for SubsonicController
Subsonic clients call this API to see what media users are playing.
If a client calls this, returning an empty list is better than an
a 404 or code 70 for non-supported route, since what a server may
return for unimplemented API routes may vary.

Ideally, this should be expanded to actually return what's actually
playing on the server.

For an example of an affected client, see SubmarinerApp/Submariner#148.