Scroll navigation pane automatically on view activation
In case there are enough playlists to make the navigation pane scrollable, the
link to the activated view may be off the view port. In such occasions, the
navigation pane is now scrolled automatically upon view activation to make the
link of the activated view visible. This is useful mostly when the view
activation happens as response to some other action than clicking the
navigation pane link (say, creating a new playlist or clicking the playing
song name on the controls pane to jump to the playing view).

refs https://github.com/owncloud/music/issues/1083