Ampache: Include only own albums in the album count of an artist
The property `albumcount` on artist-type results now includes only those
albums where the artist in question is the album artist. Any other albums
featuring this artist are ignored. The advanced search for artists
with the rule `album_count` follows the same logic.

This is how the Ample client assumes these features to work in its "Album
Artists" view. So most likely, this is also how the real Ampache server works
here.