Ampache: Support advanced search by regex also on PostgreSQL
The syntax for regular expressions is different on PostgreSQL compared
to MySQL/MariaDB. We now check the DB type from the config (i.e. from
the contents of config.php) and variate the SQL queries accordingly.

Note: Newer versions of SQLite should understand the same regex syntax
as MySQL. But the gotcha is that SQLite is not shipped with a built-in
regex engine and that should be somehow injected as an external
function. That might be impossible for us as the SQLite driver is buried
so deep under the NC/OC abstractions.

Note2: The regex patterns supported on PostgreSQL and MySQL are not
exactly the same although some basic things work the same.