Fix database migration failing on some versions of NC
It turned out that the workaround implemented in f9a95569 didn't work
after all on all versions of Nextcloud and SQLite. The reason for this
is unknown as it definitely worked on some environments.

Hence, we now fall back to the crude workaround of declaring column
`ampache_user_id` as nullable although logically it shouldn't be.