Fix the NC25+ style navigation pane highlights on NC28
The highlight of the active view was not visible on NC28.

Also, we now highlight also the hovered menu item with a background color like
the native apps of Nextcloud do in the recent versions.

TODO: The rounded background highlight is now visible also on the old NC
versions which use square layout everywhere. This doesn't look very nice.