Fix lite player attempting to play also audio files with unsupported MIME types
There was no pass-through for the method `canPlayMime` in the EmbeddedPlayer
class. In addition, the syntax used for `_.filter` was not valid at least in
our current version of Lodash.

This logic added six years ago in the commit 28509096 has apparently been
broken for the last five years, since the commit e47edf93c.