Register the playlist file tab view in the Files app also on NC28
The old API for registration is no longer available on NC28 and a new
one has to be used. The core functionality of the view is now
implemented in our stand-alone class, not inheriting any core interface.
It is then just wrapped differently depending on the cloud version used.

What doesn't currently work on NC28, is starting playback of a new list
by clicking a song name in the playlist tab view. This still works on
the older clouds, though.