Fix small layout issues on the pop-over menu of the embedded player
The highlight of the menu didn't look very nice on the recent NC
versions, probably starting from NC25. Moving the padding to an inner
element fixed this without harming the layout on OC. The padding is
needed in the first place for OC, NC applies enough padding here
natively.

Also, the "more" button opening the menu was misplaced a bit on the
recent NC versions.