Add own input field for the search within the Music app view
The new global search modal used on NC28 was not really suitable to use
as an input field for the local search within the active Music app
view. And even the "unified search" introduced a few NC major version
ago was a bit awkward to use as input for our search.

We now have our own search input field docked to the bottom of the
navigation pane, just above the Settings link.