Fix bottom-docked navigation pane items placement on mobile layout on NC25+
Using the variable --body-container-margin to adjust the position of
these items didn't work correctly because this variable gets value 0 on
the mobile layout. Meanwhile, the cloud core uses the expression
calc(var(--default-grid-baseline)*2) for similar purpose on normal
navigation pane items, and this remains the same regardless of the
window width.