Fix clear search button being misaligned on Chrome
The button was being shown off the vertical middle of the text box on Chrome
while it looked nice on Firefox. Now it looks the same on both of these
browsers.

Also, fine-tuned the horizontal placement of the clear search button and the
vertical placement of both the search and playlist name inputs.