Enable launching the sidebar playlist tab from the embedded player on NC28
The OCA.Files.Sidebar API is now used to open the sidebar instead of
mFileList.showDetailsView whenever available. On NC28, the Sidebar API is the
only way to do this but it has actually been available for more than 4 years,
starting from NC18. The benefit of the new API is that it works also in case
the target file resides outside the currently viewed folder.