Create the play URLs for playlists files within the front-end on NC28
In the previous NC28-compatibility changes, we made the backend include the
playback URLs when it parses a playlist file and returns its contents in the
REST method `/api/playlists/file/{fileId}`. This had such a drawback that it
increased the parsing time ~25% and the response size by ~50% which could make
a difference with huge playlist files. Meanwhile, we don't actually need the
playback URL for any other song except the one which is just being played.

This was now changed so that the playback URL is created on the front-end
based on the file ID whenever the playing song changes. The logic to create
the playback URL is the same as is used in the Music app proper.