Smart list filters: Style Genres and Artists like the other inputs
NC28 has changed again the border style of the input fields and the CSS
variables used for the border colors. On the other hand, the styles we
assigned for the special "chosen" input fields never quite matched the
styling of the standard input fields on OC or old NC versions.

To get consistent styling of all filter inputs on all possible host
clouds, we now copy the border style rules from standard input fields to
the chosen fields. This required quite a bit of CSS/JS-trickery because
we can't directly get or set the :hover style of an element.