Show the embedded player busy state within the player bar itself
Previously, we used to show the "in progress" spinner on the file thumbnail
in the files list. This happened when loading or importing a playlist file.
NC28 no longer provides an API to show such progress animation and we have
changed the design: The animation is now shown on the player bar, overlaid on
the album art. This design works the same regardless of the cloud version.