Fix shared folders not being shown in the folder tree view on NC28
While the externally mounted folders have their parent ID set as -1 on NC28,
the folders shared with the current user have the parent ID of the parent
folder in the share owner's file system.

Both of these cases are now handled on the front-end and the back-end-side
fix previously added for the externally mounted folders was ditched.