Make the DB column music_ampache_sessions.ampache_user_id nullable also on OC
The field was made nullable on Music 1.9.1 for Nextlcoud to work around the
limitations of SQLite. At that time, the perception was that the problem
did not affect ownCloud and the change was not applied there. Now, however,
the same problem was seen also on ownCloud 10.4 while testing the SW upgrade.