Sanitize the advanced search operators to prevent SQL injection attacks
This wasn't critical previously while advanced search was accessible
only from the Ampache API because there, all the operators were created
by our own code withing AmpacheController. But as we have added the
advanced search to the web UI, the operators now may come directly from
the REST API and cannot be trusted to be non-malicious.