Fix the background jobs not working on NC29+
The base class OC\BackgroundJob\TimedJob has been removed in NC29 in
favor of OCP\BackgorundJob\TimedJob. However, the new alternative is
not available on ownCloud and we now need to dynamically select among
these two the one which can be found.

Also, marked the app now as compatible with NC29 as it seemed to work
fine on NC29-RC1. Except for the Nextcloud  files dialog issue
https://github.com/nextcloud/server/issues/42291 which has made a
come-back, after being already fixed in NC28.0.2. [sigh]

refs https://github.com/owncloud/music/issues/1132