Make the ampache session timeout configurable for the cloud admins
The timeout can now be set with the `config.php` key
`music.ampache_session_expiry_time`. The default is 6000 seconds and
maximum is 31536000 seconds, equaling one year. The maximum has been
limited to prevent overflow especially with PostgreSQL where the expiry
timestamp (as Unix epoch time) cannot exceed 2^31-1 (which equals a
date in the year 2038).

refs https://github.com/owncloud/music/issues/1134