Fix advanced search operator "does not sound like" not working
It turned out that there is actually no operator `NOT SOUNDS LIKE` on
MySQL/MariaDB (nor is there `SOUNDS NOT LIKE`). The proper way to make
such negated query would be to use `NOT a SOUNDS LIKE b`. Breaking the
operator into two parts like this wouldn't fit our current design.

To solve this, we no longer use the `SOUNDS LIKE` operator at all but
instead formulate an equivalent statement using the `SOUNDEX` function:
`SOUNDEX(a) = `SOUNDEX(b)` or `SOUNDEX(a) != SOUNDEX(b)`.

The added benefit of this is that this is not a MySQL/MariaDB-specific
syntax but works out-of-the-box also on SQLite. Furthermore, it works
also on PostgreSQL by installing the `fuzzystrmatch` extension!