Fix advanced search numeric rules not working properly on SQLite
All the numeric parameters were being passed to the DB engine as
strings. This worked fine on MySQL/MariaDB which apparently converted
the strings to ints in expected manner. SQLite, on the other hand,
returned totally unexpected results.

Fixed by casting all the numeric parameters to integers before passing
them to the DB. This required separating the date-type operators
'before' and 'after' from the actual numeric operators as the date type
parameters must not be casted to integers.