Fix advanced search rules `album_count` and `song_count` with 0-count
The advanced search rules `album_count` and `song_count` on the artists
always omitted the artists whose respective count was 0. This happened
regardless of the operator and input used, even if those were something
like `= 0` or `< 1`.

To make this work properly, a pretty complex multi-level nested query
had to be written. I hope I never need to touch it again. The new query
has been tested to work correctly on MariaDB and SQLite.