Support advanced search regexp rules also when SQLite used
SQLite doesn't ship with an implementation for regular expression
matching but we can register a custom implementation for it which is
implemented using the PHP standard `preg_match` function.

The only difficulty on this is to gain access to the underlying PDO API
providing the function to register SQLite functions. This has to be dug
out using a private API which is different between OC and NC.