Fix a bug on track click with colliding ID with podcast/radio
The logic of handling the track click in most views didn't work
correctly in case there was currently a podcast episode or radio station
playing and the said episode/station happened to have the same ID
number as the clicked track. In that case, the current playback got
paused instead of jumping to the clicked track.

The other way around, similar situation was already handled properly in
the Podcasts and Internet radio views.