Fix every Subsonic XML API call logging an error on NC28+
Every call to the Subsonic XML API caused and error 'Undefined array
key "" at /var/www/html/lib/private/AppFramework/Http.php#128' to be
logged to the nextcloud.log.

This happened because the XmlResponse object had no response status
set. Setting this used to happen automatically, but after the Nextcloud
PR https://github.com/nextcloud/server/pull/38802, the automatic setting
has not happened without a call to the parent class constructor from
the derived class. However, we can't call the parent constructor because
it doesn't exist on ownCloud. To overcome the problem, we now set the
status code explicitly in the constructor of XmlResponse.

refs https://github.com/owncloud/music/issues/1142