Fix a glitch when opening audio file from Dashboard on NC28+
There was such a problem that audio file opened from the NC28+
Dashboard app opened using the built-in viewer instead of the embedded
player of the Music app. This happened because the embedded player was
registered a tad too late for this case.

What made this a bit tricky to fix was that, on ownCloud and older
Nextcloud versions, the file actions cannot be registered before the
page has finished loading. And we don't even know if we need to do the
old-style registration before the page has completely loaded. And
that's too late for NC28+. Hence, we need to attempt the NC28-style
registration as soon as possible and then also try the OC-style
registration later.

The described problem was seen on Chrome but not on Firefox.

refs https://github.com/owncloud/music/issues/1126