Fix Music app page loading occasionally failing on Chrome
It happens sometimes (but quite rarely at least on my tests) that all
the needed parts of the Nextcloud server core are not yet loaded at the
time when our controllers are being initialized. Previously such cases
caused the page loading to fail but now we are prepared: If the needed
parts of the core are not available immediately, then we postpone the
related initialization steps and retry after 500 ms. Attempts are
continued until when they succeed.

In the GitHub issue, the problem was reported to happen with the
reference OC.MimeType, but fixing this case reveled that there was a
similar case related to the HTML element #app-navigation-toggle.

refs https://github.com/owncloud/music/issues/1137