Ampache: Support authentication using the bearer token header
Instead of sending the session token as an URL argument, it's now
alternatively possible to deliver it using the HTTP header
`Authorization` with the content "Bearer <token>".

This required using the PHP function getallheaders() which is available
for all HTTP servers starting from PHP 7.3 but only for Apache on older
PHP versions.

refs https://github.com/owncloud/music/issues/1140