Adjust the behat tests to the changed ordering in Subsonic results
After the commit b29d63e2, the Subsonic API returns the Internet radio
stations and playlists sorted alphabetically. Previously, the order was
undefined but the behat tests assumed that the items would be returned in the
creation order. This assumption happened to be valid on MySQL but not
necessarily on all other database systems.