Ampache: Set the `has_art` property properly in the `user` action
On recent versions of Nextcloud, the avatar is always available; either
as user-provided or as a placeholder graphic. On ownCloud, on the other
hand, the avatar is available via the normal URL only if it has been
provided by the user. Otherwise, the URL returns a simple JSON content.
In those cases, the `has_art` is now returned as false.