Drop support for PHP < 7.4 and related NC/OC versions
Old PHP 7 versions have some severe security vulnerabilities and should
not really be used anymore. Also PHP 7.4 has been in end-of-life for a
while but it's still used by many and is the only version supported on
ownCloud since version 10.12.0. That is, OC still doesn't support any
PHP version 8.x and it's uncertain if it ever will.

As we drop support for the older PHP versions, it no longer makes sense
to try to support such ownCloud and Nextcloud versions which are not
compatible with PHP 7.4 or newer. This includes OC versions 10.0-10.4
and NC 13-17.