Small tweak to the albums compact layout on Chrome
On Chrome, the #app-view element is 3 pixels narrower than on Firefox,
possibly because of the space reserved for the scrollbar. The albums
compact layout has been designed so that the view could fit
four columns on 1920p display with 100% scaling, but this actually
worked only on Firefox. On Chrome, there were only three columns. This
could be fixed by reducing the column width by 1 pixel.