Show placeholder art for playlists with no proper art
In case a playlist doesn't contain any tracks which would have album
art set, the server now provides a placeholder graphic. These cases
cannot be handled easily using just front-end logic because a rather
complex DB query is needed to even know if the album covers are
available or not.