Use the EventDispatcher API to load the embedded player on Nextcloud
The EventDispatcher should actually be available also on ownCloud but the
event types we use are introduced by NC17 and NC20. The legacy events
'OCA\Files::loadAdditionalScripts' and
'OCA\Files_Sharing::loadAdditionalScripts' were removed in NC28. For now, we
use our old way of regex matching the URL on ownCloud.