Albums view: Fix some tracks reserving too little space for "Details" button
In case the track had so long name it got truncated, and the the track had
somewhat wide number (e.g. 2-23 on a multi-disk album), there was too little
space left for the "Details" button and it was difficult to click. Now the
truncation logic takes the number width into account and this work better.
The same problem and fix applies also on the Podcasts view.

Note that in the "single column" mode used on narrow displays, we have opted
to keep everything as it was. There, we allow the track name to wrap over to
multiple lines. Some overly long names are still handled there quite badly
and the "Details" button may be difficult or impossible to click. However, it
was found to be extremely difficult to fix this without creating new
problems.