Utilize EventDispatcher to register the embedded player also on ownCloud
In addition, a bit of cleanup was made on app.php. One noteworthy change is
that we now use again the plain path to music.svg instead of the hashed path
from HtmlUtil. Nextcloud was anyway using the plain music.svg (via info.xml),
so ownCloud can as well follow the suite.