On NC, don't load the embedded player on auth page nor on shared file
The typed dispatcher event BeforeTemplateRenderEvent on Nextcloud from the
Files_Sharing app tells us if the loaded view is the authentication page of
a password-protected share and if the shared node type is file or folder. On
NC, we anyway are using the embedded player only on folder shares and now we
can avoid the unnecessary loading of our module on the authentication page or
on the share page of an individual file.