Enable internal use of the Ampache API without an API key
The Ampache API should be a good fit for our Dashboard widget and now it's
possible to utilize this API for the logged-in user without setting up a
separate API session (requiring the generated API key). The API used by our
main web UI is a much poorer match. The Shiva API would be an alternative,
but it lacks support for podcasts and Internet radio, at least.