Play songs on the dashboard widget by clicking the song name
Enabling playback basically finishes the MVP for the widget and the rest of
the work is just incremental improvement.

The small optimization of using `get_indexes` to get "All tracks" had to be
ditched because we need the MIME type of the song to play it. Hence, we now
fetch the full objects with the `songs` action.