Don't use `ctype_digit` in Scanner.php
The PHP function `ctype_digit` is bad because it behaves differently before
and after PHP 8.1: Before, integer values [-128, 255] are treated as ASCII
codes and after they are treated the same as any other integer i.e. as if the
input was a string holding that number. In practice, this probably didn't
cause any visible error in the Music app but there was a warning emitted to
the log whenever the scanned year field was not a string (e.g. because it was
null).

The function `is_numeric` works more systematically and should match the need
just as well in our case of interpreting the track release year.