Use Backbone events instead of ng events in playlistService
Modified the Music app proper to utilize the Backbone events instead of the
Angular events with the playlistService. This is in preparation of making the
service reusable also for the dashboard widget.

There was previously also such a design smell that the view controllers
implicitly assumed that the playlistService uses the Angular $rootScope
events under the hood and subscribed to those directly, bypassing the
playlistService interface.