Fix cover art updating not always working correctly with PostgreSQL
With PostgreSQL, uploading an image file caused an unhandled exception
Doctrine\DBAL\Exception\DriverException in case the target folder contained
songs of more than one album. Hence, the uploaded image wasn't immediately
used as a cover art for any album or artist. For album covers, the image was
pulled later by the background task but for artist cover, the file could be
detected only by running a new library scan.

The root cause was that the way used to handle the parameters of the SQL query
wasn't compatible with PgSQL.

refs https://github.com/owncloud/music/issues/1164