In dashboard widget "Album artists" mode, allow viewing "all albums"
Enable listing the tracks from all the albums of the selected album
artist. There wasn't such action readily available in the Ampache API, and
hence, this was added as a small extension to the advanced_search action.
We could have used the existing advanced_search rule "album_artist" on
songs but that would not work in the case of the "Unknown artist" (which
uses a localized title). The new proprietary rule "album_artist_id" avoids
this problem.

The "Album artists" mode of the widget was changed also so that it now
shows only the own albums of the artist and not the albums where the
artist is merely performing some track. This is better in line with the
fact that the mode only list those artists who have own albums.