Ampache API: Use URL as title for nameless radio stations
The stream url is the only mandatory field on the internet radio stations;
use this url as the station title if there's no name set. The nameless radio
stations became an issue in the dashboard widget but the problem was fixed on
the Ampache API level. We already added earlier such fallback logic to the
Subsonic API and it's good to keep these APIs in sync where possible.