Handle "previous" button similarly on dashboard widget as elsewhere
The "previous" button now seeks to the beginning of the current track on
first click. Only on second click it jumps to the previous track (or when
clicked within 2 seconds from the song start).

This is the same behavior as was already followed on the Files embedded
player and in the Music app proper.