Show proper progress on mediaSession also with Aurora.js
At least recent versions of Chrome now show a progress bar in the
mediaSession popup. As we didn't previously tell the mediaSession anything
about progress, Chrome took the progress info directly from the Audio
element. In case we were playing the audio using Aurora.js, this meant that
the progress came from the dummy audio element introduced in the commit
88d7a433.