Subsonic: Allow authentication using `apiKey`
This is the OpenSubsonic extension
[`apiKeyAuthentication`](https://opensubsonic.netlify.app/docs/extensions/apikeyauth/).

Basically, we have been using "API key" from the day one because we don't
allow any user-given password, but only generated keys. But now this
generated key may be passed using the dedicated argument `apiKey`. In this
case, the user name argument `u` must not be given.

The OpenSubsonic extension includes also the new method `tokenInfo` which may
be used by the client to obtain the user name when it knows only the API key.