Ampache: Use the more reader-friendly format for the API version
The API version is now given, for example, as 6.6.1 instead of 661000. As an
exception, if the client requests a specific version using the legacy format,
then the version is reported using this same version. This is needed for
backward-compatibility (*) and mimics the behavior of the Ampache server
proper.

Also, the version number returned in the property `version` of the action
`ping` is now the same effective version number as the one returned in the
property `api` of the `handshake`. Previously, we were returning here the
highest supported API version but this didn't match the behavior of the
Ampache server proper.

(*) At least the Kodi Ampache plugin breaks down if the server responds using
the new API version format.